Summer lovin' refactorization
* Add multiple boto profile management
* Add data validators
* Add personalized exceptions
* Update all parsers
* Remove obsolete code
* Update cfg file
* Fix a ton of bugs
* Update Dockerfile
* Simplify code in in almost all classes
* Update log file path

+++ src/utils.py
@@ -18,12 +18,16 @@ import base64
 import json
 import os
 import re
-import uuid
-import functools
 import subprocess
 import tarfile
-from botocore.exceptions import ClientError
-import src.exceptions as scar_excp
+import tempfile
+import uuid
+
+def join_paths(*paths):
+    return os.path.join(*paths)
+
+def get_temp_dir():
+    return tempfile.gettempdir()
 
 def lazy_property(func):
     ''' A decorator that makes a property lazy-evaluated.'''
@@ -35,33 +39,6 @@ def lazy_property(func):
             setattr(self, attr_name, func(self))
         return getattr(self, attr_name)
     return _lazy_property
-
-def exception(logger):
-    '''
-    A decorator that wraps the passed in function and logs exceptions
-    @param logger: The logging object
-    '''
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            try:
-                return func(*args, **kwargs)
-            except ClientError as ce:
-                print("There was an exception in {0}".format(func.__name__))
-                print(ce.response['Error']['Message'])
-                logger.exception(ce)
-            except scar_excp.ScarError as se:
-                #print("There was an exception in {0}".format(func.__name__))
-                print(se.args[0])
-                logger.exception(se)
-                raise        
-            except Exception as ex:
-                print("There was an unmanaged exception in {0}".format(func.__name__))
-                logger.exception(ex)
-                # re-raise the exception
-                raise
-        return wrapper
-    return decorator
 
 def find_expression(string_to_search, rgx_pattern):
     '''Returns the first group that matches the rgx_pattern in the string_to_search'''
@@ -80,9 +57,6 @@ def utf8_to_base64_string(value):
 def dict_to_base64_string(value):
     return base64.b64encode(json.dumps(value)).decode("utf-8")
 
-def print_json(value):
-    print(json.dumps(value))
-    
 def divide_list_in_chunks(elements, chunk_size):
     """Yield successive n-sized chunks from th elements list."""
     if len(elements) == 0:
@@ -93,22 +67,23 @@ def divide_list_in_chunks(elements, chunk_size):
 def get_random_uuid4_str():
     return str(uuid.uuid4())
 
-def has_dict_prop_value(dictionary, value):
-    return (value in dictionary) and dictionary[value] and (dictionary[value] != "")
-
-def load_json_file(file_path):
-    if os.path.isfile(file_path):
-        with open(file_path) as f:
-            return json.load(f)
-        
 def merge_dicts(d1, d2):
+    '''
+    Merge 'd1' and 'd2' dicts into 'd1'.
+    'd2' has precedence over 'd1'
+    '''
     for k,v in d2.items():
         if v:
-            d1[k] = v
+            if k not in d1:
+                d1[k] = v
+            elif type(v) is dict:
+                d1[k] = merge_dicts(d1[k], v)
+            elif type(v) is list:
+                d1[k] += v
     return d1
 
-def check_key_in_dictionary(key, dictionary):
-    return (key in dictionary) and dictionary[key] and dictionary[key] != ""
+def is_value_in_dict(dictionary, value):
+    return value in dictionary and dictionary[value]
 
 def get_tree_size(path):
     """Return total size of files in given path and subdirs."""
@@ -144,14 +119,15 @@ def read_file(file_path, mode="r"):
         return content_file.read()
     
 def delete_file(path):
-    os.remove(path)
+    if os.path.isfile(path):
+        os.remove(path)
     
 def create_tar_gz(files_to_archive, destination_tar_path):
     with tarfile.open(destination_tar_path, "w:gz") as tar:
         for file_path in files_to_archive:
             tar.add(file_path, arcname=os.path.basename(file_path))
     return destination_tar_path
-        
+         
 def extract_tar_gz(tar_path, destination_path):
     with tarfile.open(tar_path, "r:gz") as tar:
         tar.extractall(path=destination_path)
@@ -167,13 +143,23 @@ def execute_command_and_return_output(command):
     return subprocess.check_output(command).decode("utf-8")
 
 def is_variable_in_environment(variable):
-    return check_key_in_dictionary(variable, os.environ)
+    return is_value_in_dict(os.environ, variable)
 
 def set_environment_variable(key, variable):
-    if key and variable and key != "" and variable != "":
+    if key and variable:
         os.environ[key] = variable
 
 def get_environment_variable(variable):
-    if check_key_in_dictionary(variable, os.environ):
+    if is_variable_in_environment(variable):
         return os.environ[variable]
 
+def parse_arg_list(arg_keys, cmd_args):
+    result = {}
+    for key in arg_keys:
+        if type(key) is tuple:
+            if key[0] in cmd_args and cmd_args[key[0]]:
+                result[key[1]] = cmd_args[key[0]]
+        else:
+            if key in cmd_args and cmd_args[key]:
+                result[key] = cmd_args[key]
+    return result
